Move Team Tournament endpoints
