+++ HandballBackend/Controllers/TeamsController.cs
@@ -1,9 +1,11 @@
+using HandballBackend.Authentication;
 using HandballBackend.Utils;
 using HandballBackend.Database;
 using HandballBackend.Database.Models;
 using HandballBackend.Database.SendableTypes;
 using HandballBackend.EndpointHelpers;
 using HandballBackend.ErrorTypes;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -217,5 +219,191 @@ public class TeamsController : ControllerBase {
             Pooled = poolOne is not null,
             Tournament = returnTournament ? tournament!.ToSendableData() : null
         };
+    }
+
+
+    public class AddTeamRequest {
+        public required string TournamentSearchableName { get; set; }
+        public string? TeamName { get; set; }
+        public string? CaptainName { get; set; }
+        public string? NonCaptainName { get; set; }
+        public string? SubstituteName { get; set; }
+    }
+
+
+    public class AddTeamResponse {
+        public required TeamData Team { get; set; }
+    }
+
+    [HttpPost("/addTeamToTournament")]
+    [Authorize(Policy = Policies.IsAdmin)]
+    public async Task<ActionResult<AddTeamResponse>> AddTeamToTournament(
+        [FromBody] AddTeamRequest request) {
+        var db = new HandballContext();
+        var tournament = db.Tournaments
+            .FirstOrDefault(a => a.SearchableName == request.TournamentSearchableName);
+        if (tournament is null) {
+            return NotFound("Invalid Tournament");
+        }
+
+        if (tournament.Started) {
+            return NotFound("Tournament has already started!");
+        }
+
+        var team = await db.Teams.IncludeRelevant().Include(team => team.TournamentTeams)
+            .FirstOrDefaultAsync(t => t.Name == request.TeamName);
+        if (team is not null && (request.CaptainName is not null || request.NonCaptainName is not null ||
+                                 request.SubstituteName is not null)) {
+            return BadRequest("This Team already exists; you may not provide players");
+        }
+
+
+        if (team is null) {
+            var playerIds = ((string?[]) [request.CaptainName, request.NonCaptainName, request.SubstituteName])
+                .Where(s => !string.IsNullOrEmpty(s))
+                .Select(a => db.People.FirstOrDefault(p => p.Name == a)?.Id)
+                .ToList();
+            while (playerIds.Count < 3) {
+                playerIds.Add(null);
+            }
+
+            var maybeTeam = db.Teams.IncludeRelevant().Include(t => t.TournamentTeams).FirstOrDefault(t =>
+                // Both players must be in one of the roles
+                (playerIds.Contains(t.CaptainId ?? null) &&
+                 playerIds.Contains(t.NonCaptainId ?? null) &&
+                 playerIds.Contains(t.SubstituteId ?? null)) &&
+
+                // Count of non-null player references should be exactly 2
+                ((t.CaptainId.HasValue ? 1 : 0) +
+                    (t.NonCaptainId.HasValue ? 1 : 0) +
+                    (t.SubstituteId.HasValue ? 1 : 0) == playerIds.Count(a => a.HasValue))
+            );
+            if (maybeTeam == null) {
+                team = new Team {
+                    CaptainId = playerIds![0],
+                    NonCaptainId = playerIds[1],
+                    SubstituteId = playerIds[2],
+                    Name = request.TeamName!,
+                    SearchableName = Utilities.ToSearchable(request.TeamName!)
+                };
+                await db.Teams.AddAsync(team);
+                await db.SaveChangesAsync();
+            } else {
+                team = maybeTeam;
+            }
+        }
+
+        if (team.TournamentTeams.Any(tt => tt.TournamentId == tournament.Id)) {
+            return BadRequest("That team is already in this tournament!");
+        }
+
+        await db.TournamentTeams.AddAsync(new TournamentTeam {
+            TournamentId = tournament.Id,
+            TeamId = team.Id,
+            Name = request.TeamName == null || request.TeamName == team.Name ? null : request.TeamName,
+        });
+
+
+        await db.SaveChangesAsync();
+        return Ok(new AddTeamResponse {
+            Team = team.ToSendableData()
+        });
+    }
+
+
+    public class UpdateTeamRequest {
+        public required string TournamentSearchableName { get; set; }
+        public required string TeamSearchableName { get; set; }
+        public string? NewName { get; set; }
+        public string? NewColor { get; set; }
+    }
+
+    public class UpdateTeamResponse {
+        public required TeamData Team { get; set; }
+    }
+
+    [HttpPatch("updateTeamForTournament")]
+    [TournamentAuthorize(PermissionType.UmpireManager)]
+    public async Task<ActionResult<UpdateTeamResponse>> UpdateTeamForTournament(
+        [FromBody] UpdateTeamRequest request) {
+        var db = new HandballContext();
+        var tournament = await db.Tournaments
+            .FirstOrDefaultAsync(a => a.SearchableName == request.TournamentSearchableName);
+        if (tournament is null) {
+            return NotFound("Invalid Tournament");
+        }
+
+        if (tournament.Started) {
+            return NotFound("Tournament has already started!");
+        }
+
+        var team = await db.Teams.IncludeRelevant().Include(team => team.TournamentTeams)
+            .SingleAsync(team => team.SearchableName == request.TeamSearchableName);
+
+        if (team.TournamentTeams.All(tt => tt.TournamentId != tournament.Id)) {
+            return BadRequest("Team not in tournament!");
+        }
+
+        var tournamentTeam = team.TournamentTeams.Single(tt => tt.TournamentId == tournament.Id);
+        if (team.TournamentTeams.Count(tt => tt.Id != 1) == 1) {
+            if (request.NewName != null) {
+                team.Name = request.NewName;
+                team.SearchableName = Utilities.ToSearchable(request.NewName);
+            }
+
+            if (request.NewColor != null) {
+                team.TeamColor = request.NewColor;
+            }
+        } else {
+            if (request.NewName != null) {
+                tournamentTeam.Name = request.NewName;
+            }
+
+            if (request.NewColor != null) {
+                tournamentTeam.TeamColor = request.NewColor;
+            }
+        }
+
+
+        await db.SaveChangesAsync();
+        return Ok(new UpdateTeamResponse {
+            Team = tournamentTeam.ToSendableData()
+        });
+    }
+
+    public class RemoveTeamRequest {
+        public string? TeamSearchableName { get; set; }
+        public string? TournamentSearchableName { get; set; }
+    }
+
+    [HttpDelete("removeTeamFromTournament")]
+    [TournamentAuthorize(PermissionType.UmpireManager)]
+    public async Task<ActionResult> RemoveTeamFromTournament([FromBody] RemoveTeamRequest request) {
+        var db = new HandballContext();
+        var tournament = await db.Tournaments
+            .FirstOrDefaultAsync(a => a.SearchableName == request.TournamentSearchableName);
+        if (tournament is null) {
+            return NotFound("Invalid Tournament");
+        }
+
+        if (tournament.Started) {
+            return NotFound("Tournament has already started!");
+        }
+
+        var team = await db.Teams.Include(team => team.TournamentTeams)
+            .SingleAsync(t => t.SearchableName == request.TeamSearchableName);
+        var deleteTeam = team.TournamentTeams.Count(tt => tt.TournamentId != 1) < 1;
+
+        var tournamentTeam = team.TournamentTeams.Single(tt => tt.TournamentId == tournament.Id);
+        db.TournamentTeams.Remove(tournamentTeam);
+        if (deleteTeam) {
+            db.Teams.Remove(team);
+        }
+
+        db.TournamentTeams.Remove(tournamentTeam);
+        await db.SaveChangesAsync();
+
+
+        return Ok();
     }
 }
